Add ticket comments with email notifications #125
- Create TicketComment entity with author, role, and timestamp
- Add TicketCommentRepository and TicketCommentService
- Add GET/POST endpoints for ticket comments
- Send async email notifications on new comments
- Add generic send() method to EmailService
- Notify tenant when admin replies, notify admin when tenant replies

+++ src/main/java/za/co/tms/controller/TicketController.java
@@ -43,12 +43,12 @@ public class TicketController {
 	}
 
 	@GetMapping(path="/find/by/{id}")
-	public Ticket retrieveHelpdeskTicketById(@PathVariable int id) {
+	public Ticket retrieveHelpdeskTicketById(@PathVariable Long id) {
 		return ticketService.findHelpdeskTicketById(id);
 	}
 
 	@DeleteMapping(path="/delete/{id}")
-	public ResponseEntity<Void> deleteHelpdeskTicketManagementById(@PathVariable int id) {
+	public ResponseEntity<Void> deleteHelpdeskTicketManagementById(@PathVariable Long id) {
 		ticketService.deleteHelpdeskTicketById(id);
 		return ResponseEntity.noContent().build();
 	}
